Implement liveness-aware CallTraceStorage with dropped trace handling
Major architectural changes:
- Replace monolithic CallTraceStorage with double-buffered hash table design
- Add CallTraceHashTable with lock-free concurrent access and instance-based trace IDs
- Implement liveness tracking system to preserve active traces across JFR dumps
- Add dropped trace handling for lock contention with proper JFR integration

Key features:
- 64-bit trace IDs combining instance ID and slot for collision avoidance
- Split-lock strategy minimizing exclusive lock time during trace collection
- Platform-specific ASGCT_CallFrame alignment using LP64_ONLY macro
- Comprehensive test coverage including contention and liveness scenarios

+++ ddprof-lib/src/main/cpp/objectSampler.cpp
@@ -1,18 +1,7 @@
 /*
  * Copyright 2022 Andrei Pangin
- * Copyright 2022, 2023 Datadog, Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Copyright 2022, 2025, Datadog, Inc.
+ * SPDX-License-Identifier: Apache-2.0
  */
 
 #include <algorithm>
@@ -68,7 +57,7 @@ void ObjectSampler::recordAllocation(jvmtiEnv *jvmti, JNIEnv *jni,
     event._id = id;
   }
 
-  u32 call_trace_id = 0;
+  u64 call_trace_id = 0;
   // we do record the details and stacktraces only for when recording
   // allocations or liveness
   if (_record_allocations || _record_liveness) {
